Another commit
Added chackpoint system
Added the UI representation of the the ammo
The player saving thing is suggest we talk over cause it can softlock the player some times.
And I added the design for the StartLevel, END is where i suggest we add the transition.

Hope everything is fine.
Good night!!!